requirements.txt
openai
transformers
flask
groq
python-dotenv